server/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
httpx>=0.27.0,<1.0.0
cryptography>=43.0.0,<46.0.0
psutil>=6.0.0,<8.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0

